requirements.txt
aiofiles==25.1.0
Flask==3.1.3
flask_cors==6.0.2
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.24.0
langchain==1.2.10
langchain_classic==1.0.2
langchain_community==0.4.1
langchain_core==1.2.17
langchain_huggingface==1.2.1
langchain_openai==1.1.10
langchain_pinecone==0.2.13
langchain_text_splitters==1.1.1
langchain_weaviate==0.0.6
numpy==2.4.2
openai==2.26.0
pytest==9.0.2
python-dotenv==1.2.2
weaviate_client==4.20.3
WTForms==3.2.1
